requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
packaging==24.2
pillow==11.1.0
PyJWT==2.9.0
sqlparse==0.5.3
typing_extensions==4.13.2
whitenoise==6.9.0
psycopg2-binary==2.9.9
python-dotenv==1.1.0
woocommerce==3.0.0